Support multiple types with the same id in the replication.
Merkle tree replication must support multiple types with the same id. Types represent thumbs of a original file.

A block has id and a list of elements (types). The merkle tree operates ranges of original id, but the hash of a leaf node includes all types of the orignal id (in the given partition). So, even if, the set of original elements are identical, but types are different, hashes will be different as well.